requirements.txt
PySide6>=6.7
pymodbus>=3.6,<4.1
python-snap7>=2.0